+++ drivers/serial.c
@@ -26,7 +26,7 @@ void serial_put(char a)
 
 void serial_write(char* str)
 {
-	for (size_t i = 0; i < strlen(str); i++)
+	for (int i = 0; i < strlen(str); i++)
 		serial_put(str[i]);
 }
