+++ include/terminal.h
@@ -6,7 +6,7 @@
 void init_terminal(void);
 void terminal_setcolor(uint8_t color);
 void twrite(const char* data);
-void terminal_write(const char* data, size_t size);
+void terminal_write(const char* data, int size);
 void draw_mem_usage(int used);
 void terminal_putchar(char c);
 int32_t twritef(char* fmt, ...);
